Get a summary of available information about a lake or lakes from the UK CEH Lakes Portal

Produces a summary of available information about a lake or lakes from the UK CEH Lakes Portal https://uklakes.ceh.ac.uk/ by responsibly webscraping lake information pages.

Usage

get_lake_info(...)

Arguments

...

A lake wbid number or series of lake wbid numbers to be searched. Individual values should be separated by commas. R-type sequences can also be passed in e.g. 34:48, or alternatively a vector of values or df column. See examples.

Value

A data frame containing the available information about the specified lake or lakes. Biology, land cover and water quality information on the Lakes Portal webpages are not included, but all other information is provided. For details of the information provided, see the UK CEH Lake Portal website (https://uklakes.ceh.ac.uk/).

Examples

if (FALSE) { # \dontrun{
# get information for Loch Lomond and Loch Katrine
get_lake_info(24447, 24531)

# get lake information for a set of lakes including a sequence
get_lake_info(4, 5, 6, 34:37)

# can also pass in the wbid column from the output of search_lakes
tarn_list <- search_lakes("Tarn")
# just retrieve the first 5 lakes
get_lake_info(tarn_list$wbid[1:5])
} # }