Search database for lakes within a set buffer distance of a point

Searches the listing of UK CEH Lakes Portal lake sites to find lakes that have their centroid within a specified buffer distance from the point given.

Coordinates are specified as British National Grid X and Y and are checked to ensure that they are within the required ranges.

Usage

search_buffer(xcoord, ycoord, buffer_distance)

Arguments

xcoord

The x coordinate, in British National Grid, of the point to base the buffer on.

ycoord

The y coordinate, in British National Grid, of the point to base the buffer on.

buffer_distance

The buffer distance (radius) to be searched around the point in metres.

Value

A data frame containing the details (Name, wbid, Country) of all the lakes that have their centroid within the specified buffer distance of the point as well as the distance (in metres) from each lake to the point.

Examples

# search for sites within 1km of the point 315231, 642134
search_buffer(335792, 501379, 2000)
#>                    Name  wbid Country  Distance
#> 26538           unnamed 29237 England 1680.8822
#> 26562    High Crag Tarn 29264 England  807.6664
#> 26567      Blelham Tarn 29270 England 1208.7452
#> 26568 Blelham Fish Pond 29271 England  864.0978