Search database of site names

Searches the listing of UK CEH Lakes Portal sites to find names that contain the string provided.

The search is done on an internal copy of the UK lakes listing rather than connecting to the UKCEH Lakes Portal site.

Usage

search_lakes(string = NULL)

Arguments

string

The search string to be matched (case-sensitive). Will match whole or partial strings in the column values. #'

Value

A data frame containing the details (Name & wbid) of all the sites that match the search string (full or partial matches).

Examples

# search for sites containing "Tarn" in the name
search_lakes("Tarn")
#>                                Name  wbid
#> 1                       Tarnie Loch 23031
#> 2                       Talkin Tarn 28386
#> 3                      Tindale Tarn 28395
#> 4                       Martin Tarn 28561
#> 5                         Tarns Dub 28651
#> 6                  Greencastle Tarn 28736
#> 7                       Little Tarn 28827
#> 8                   Gillcambon Tarn 28838
#> 9                     Bowscale Tarn 28854
#> 10                         Tarn Dub 28899
#> 11               Great Rundale Tarn 28901
#> 12                      Scales Tarn 28905
#> 13                     Seamore Tarn 28911
#> 14              Little Rundale Tarn 28916
#> 15                       Tewet Tarn 28962
#> 16                   Mockerkin Tarn 28966
#> 17                   High Nook Tarn 29008
#> 18                    Floutern Tarn 29045
#> 19                  Watendlath Tarn 29061
#> 20                   Bleaberry Tarn 29081
#> 21              Red Tarn, Helvellyn 29083
#> 22                       Angle Tarn 29093
#> 23                        Dock Tarn 29094
#> 24                        Blea Tarn 29097
#> 25                      Harrop Tarn 29107
#> 26                   Blackbeck Tarn 29121
#> 27                   Tarn at Leaves 29126
#> 28                   Grisedale Tarn 29129
#> 29                     Greenup Tarn 29145
#> 30                       Scoat Tarn 29153
#> 31                     Styhead Tarn 29157
#> 32                         Low Tarn 29162
#> 33                  Sprinkling Tarn 29163
#> 34                    Easedale Tarn 29166
#> 35                      Codale Tarn 29167
#> 36                      Alcock Tarn 29176
#> 37                     Stickle Tarn 29177
#> 38                   Sunbiggin Tarn 29178
#> 39                       Angle Tarn 29179
#> 40                   Greendale Tarn 29181
#> 41                    Brownber Tarn 29192
#> 42                  Braystones Tarn 29200
#> 43                    Lingmoor Tarn 29210
#> 44                    Burnmoor Tarn 29215
#> 45                    Lingmore Tarn 29217
#> 46                        Blea Tarn 29218
#> 47                   Loughrigg Tarn 29219
#> 48                     Woodhow Tarn 29221
#> 49    Red Tarn or Red Tarn, Wrynose 29226
#> 50                   Greenside Tarn 29227
#> 51                  Flass Tarn West 29229
#> 52             Little Langdale Tarn 29231
#> 53                        Tarn Sike 29234
#> 54                       Stony Tarn 29246
#> 55                    Birkdale Tarn 29250
#> 56                         Eel Tarn 29252
#> 57                High Arnside Tarn 29262
#> 58                       Sinay Tarn 29263
#> 59                   High Crag Tarn 29264
#> 60                        Blea Tarn 29265
#> 61                     Blelham Tarn 29270
#> 62                    Parkgate Tarn 29272
#> 63                    Yew Tree Tarn 29274
#> 64                        Tarn Hows 29275
#> 65               Eskdale Green Tarn 29277
#> 66  Gurnal Dubs or Gurnal Dubs Tarn 29289
#> 67                  Seathwaite Tarn 29290
#> 68                      Potter Tarn 29300
#> 69                        Sand Tarn 29301
#> 70                     Hodsons Tarn 29312
#> 71                    Whinfell Tarn 29315
#> 72                    Grayrigg Tarn 29317
#> 73           Claife, Robinsons Tarn 29318
#> 74                    Wise Een Tarn 29327
#> 75               Claife, Scale Tarn 29330
#> 76                Borwick Fold Tarn 29333
#> 77                  Three Dubs Tarn 29334
#> 78                School Knott Tarn 29337
#> 79                  Black Moss Tarn 29342
#> 80                 Moss Eccles Tarn 29348
#> 81                       Blind Tarn 29351
#> 82                  Skelsmergh Tarn 29353
#> 83                      Crook Tarns 29377
#> 84                   Whitaside Tarn 29378
#> 85                Summer Lodge Tarn 29379
#> 86                    Out Dubs Tarn 29381
#> 87           Barrow Plantation Tarn 29387
#> 88                       Knipe Tarn 29388
#> 89                   Grizedale Tarn 29390
#> 90                    Cunswick Tarn 29394
#> 91             West Baugh Fell Tarn 29395
#> 92                  Cotter End Tarn 29396
#> 93                     Boltons Tarn 29400
#> 94            Fisher Tarn Reservoir 29403
#> 95                       East Tarns 29422
#> 96                       East Tarns 29423
#> 97                       East Tarns 29424
#> 98                       East Tarns 29425
#> 99                      Locker Tarn 29429
#> 100                 Green Hows Tarn 29441
#> 101                     Beacon Tarn 29446
#> 102           Great Green Hows Tarn 29450
#> 103                      Allan Tarn 29461
#> 104                    Eskrigg Tarn 29465
#> 105                  Grizedale Tarn 29467
#> 106             Widdale Little Tarn 29473
#> 107                 Tarneybank Tarn 29474
#> 108                    Sow How Tarn 29476
#> 109              Widdale Great Tarn 29478
#> 110                       Hood Tarn 29480
#> 111                   Boretree Tarn 29481
#> 112                   Barfield Tarn 29484
#> 113                     Helton Tarn 29515
#> 114                 Fleet Moss Tarn 29540
#> 115                  Whernside Tarn 29541
#> 116                  Whernside Tarn 29543
#> 117                  Tarnhouse Tarn 29544
#> 118                  Whernside Tarn 29550
#> 119                    Bigland Tarn 29552
#> 120                  Terrybank Tarn 29556
#> 121                Oughtershaw Tarn 29561
#> 122                South Grain Tarn 29568
#> 123              Middle Tongue Tarn 29572
#> 124                       Cray Tarn 29576
#> 125                 Knottallow Tarn 29586
#> 126                      Birks Tarn 29665
#> 127                    Urswick Tarn 29712
#> 128                   Standing Tarn 29720
#> 129                       Mere Tarn 29756
#> 130             Fountains Fell Tarn 29765
#> 131                    Priests Tarn 29804
#> 132                     Malham Tarn 29844
#> 133                       Blow Tarn 29947
#> 134             Blea Tarn Reservoir 30001
#> 135                     Eshton Tarn 30020
#> 136                   Brennand Tarn 30083
#> 137                        The Tarn 30194
#> 138                     Redcar Tarn 30366
#> 139                     Yeadon Tarn 30381
#> 140                        The Tarn 43758