Search database of site names

Searches the listing of UK CEH Lakes Portal lake sites to find names that contain the string provided.

The search is done on an internal copy of the UK lakes listing rather than connecting to the UKCEH Lakes Portal site.

Usage

search_lakes(string = NULL)

Arguments

string

The search string to be matched (case-sensitive). Will match whole or partial strings in the column values. #'

Value

A data frame containing the details (Name & wbid) of all the sites that match the search string (full or partial matches).

Examples

# search for sites containing "mere" in the name
search_lakes("mere")
#>                          Name  wbid
#> 1                   Thirlmere 29021
#> 2                  Buttermere 29052
#> 3          Kentmere Reservoir 29173
#> 4                    Grasmere 29184
#> 5                  Windermere 29233
#> 6                     Kitmere 29505
#> 7                 Wyndhammere 29507
#> 8                Cleveleymere 30159
#> 9             Thorpemere Pond 31569
#> 10 The Mere or Ellesmere Mere 34990
#> 11                   Fenemere 35620
#> 12              Langmere Lake 35794
#> 13                Bomere Pool 36544
#> 14               Shomere Pool 36578
#> 15                    Bosmere 38980
#> 16                   Fowlmere 39343
#> 17                  Southmere 42238
#> 18             Englemere Pond 42757
#> 19  Bolder Mere or Boldermere 43218
#> 20                  Sandymere 44541
#> 21                Burton mere 46071